Implemented rearranging and removing counter bar items through drag and drop
